+++ lib/src/in_app_webview/in_app_webview_controller.dart
@@ -55,6 +55,7 @@ class InAppWebViewController {
       HashMap<String, JavaScriptHandlerCallback>();
   List<UserScript> _userScripts = [];
   Set<String> _webMessageListenerObjNames = Set();
+  Map<String, ScriptHtmlTagAttributes> _injectedScriptsFromURL = {};
 
   // ignore: unused_field
   dynamic _id;
@@ -102,6 +103,7 @@ class InAppWebViewController {
   Future<dynamic> handleMethod(MethodCall call) async {
     switch (call.method) {
       case "onLoadStart":
+        _injectedScriptsFromURL.clear();
         if ((_webview != null && _webview!.onLoadStart != null) ||
             _inAppBrowser != null) {
           String? url = call.arguments["url"];
@@ -865,6 +867,22 @@ class InAppWebViewController {
               _webview!.onWindowBlur!(this);
             else if (_inAppBrowser != null) _inAppBrowser!.onWindowBlur();
             return null;
+          case "onInjectedScriptLoaded":
+            String id = args[0];
+            var onLoadCallback = _injectedScriptsFromURL[id]?.onLoad;
+            if ((_webview != null || _inAppBrowser != null) &&
+                onLoadCallback != null) {
+              onLoadCallback();
+            }
+            return null;
+          case "onInjectedScriptError":
+            String id = args[0];
+            var onErrorCallback = _injectedScriptsFromURL[id]?.onError;
+            if ((_webview != null || _inAppBrowser != null) &&
+                onErrorCallback != null) {
+              onErrorCallback();
+            }
+            return null;
         }
 
         if (javaScriptHandlersMap.containsKey(handlerName)) {
@@ -1372,6 +1390,10 @@ class InAppWebViewController {
       {required Uri urlFile,
       ScriptHtmlTagAttributes? scriptHtmlTagAttributes}) async {
     assert(urlFile.toString().isNotEmpty);
+    var id = scriptHtmlTagAttributes?.id;
+    if (scriptHtmlTagAttributes != null && id != null) {
+      _injectedScriptsFromURL[id] = scriptHtmlTagAttributes;
+    }
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('urlFile', () => urlFile.toString());
     args.putIfAbsent(
@@ -1385,10 +1407,10 @@ class InAppWebViewController {
   ///because, in these events, the [WebView] is not ready to handle it yet.
   ///Instead, you should call this method, for example, inside the [WebView.onLoadStop] event or in any other events
   ///where you know the page is ready "enough".
-  Future<void> injectJavascriptFileFromAsset(
+  Future<dynamic> injectJavascriptFileFromAsset(
       {required String assetFilePath}) async {
     String source = await rootBundle.loadString(assetFilePath);
-    await evaluateJavascript(source: source);
+    return await evaluateJavascript(source: source);
   }
 
   ///Injects CSS into the WebView.
@@ -2044,13 +2066,13 @@ class InAppWebViewController {
   ///[functionBody] is the JavaScript string to use as the function body.
   ///This method treats the string as an anonymous JavaScript function body and calls it with the named arguments in the arguments parameter.
   ///
-  ///[arguments] is a dictionary of the arguments to pass to the function call.
-  ///Each key in the dictionary corresponds to the name of an argument in the [functionBody] string,
+  ///[arguments] is a `Map` of the arguments to pass to the function call.
+  ///Each key in the `Map` corresponds to the name of an argument in the [functionBody] string,
   ///and the value of that key is the value to use during the evaluation of the code.
   ///Supported value types can be found in the official Flutter docs:
   ///[Platform channel data types support and codecs](https://flutter.dev/docs/development/platform-integration/platform-channels#codec),
   ///except for [Uint8List], [Int32List], [Int64List], and [Float64List] that should be converted into a [List].
-  ///All items in an array or dictionary must also be one of the supported types.
+  ///All items in a `List` or `Map` must also be one of the supported types.
   ///
   ///[contentWorld], on iOS, it represents the namespace in which to evaluate the JavaScript [source] code.
   ///Instead, on Android, it will run the [source] code into an iframe.
@@ -2058,6 +2080,11 @@ class InAppWebViewController {
   ///Those changes remain visible to all scripts, regardless of which content world you specify.
   ///For more information about content worlds, see [ContentWorld].
   ///Available on iOS 14.0+.
+  ///
+  ///**NOTE**: This method shouldn't be called in the [WebView.onWebViewCreated] or [WebView.onLoadStart] events,
+  ///because, in these events, the [WebView] is not ready to handle it yet.
+  ///Instead, you should call this method, for example, inside the [WebView.onLoadStop] event or in any other events
+  ///where you know the page is ready "enough".
   ///
   ///**NOTE for iOS**: available only on iOS 10.3+.
   ///